Do not use it yet
This is a major update there
I need to test things properly.
This will sync our build with main repo. (fast PR fetch in and out)
For now in this one there is quite a lot of change, I will list them before doing a PR.

+++ ZAPDTR/ZAPD/ZResource.cpp
@@ -288,7 +288,7 @@ void ZResource::GetSourceOutputCode([[maybe_unused]] const std::string& prefix)
 			ZArray* arr = (ZArray*)this;
 			if (arr->resList[0]->GetResourceType() == ZResourceType::Vertex)
 			{
-				for (size_t i = 0; i < arr->resList.size(); i++)
+				for (int i = 0; i < arr->resList.size(); i++)
 				{
 					ZVtx* vtx = (ZVtx*)arr->resList[i];
 					decl->vertexHack.push_back(vtx);
